fixed server part(create implements classes for this task)
