+++ server/src/main/java/com/lunchtime/implementation/DishServiseImplement.java
@@ -1,0 +1,30 @@
+package com.lunchtime.implementation;
+
+import com.lunchtime.models.Dish;
+import com.lunchtime.repository.DishRepository;
+import com.lunchtime.service.DishService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import java.util.Optional;
+
+@Service
+public class DishServiseImplement implements DishService {
+    private final DishRepository dishRepository;
+
+    public DishServiseImplement(DishRepository dishRepository) {
+        this.dishRepository = dishRepository;
+    }
+
+    public Dish save(Dish dish) {
+        return dishRepository.save(dish);
+    }
+
+    public Page<Dish> findAll(Pageable pageable) {
+        return dishRepository.findAll(pageable);
+    }
+
+    public Optional<Dish> findById(Long id) {
+        return dishRepository.findById(id);
+    }
+}
