+++ server/src/main/java/com/lunchtime/service/MenuItemDishService.java
@@ -4,29 +4,16 @@ import java.util.Optional;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
 
 import com.lunchtime.models.MenuItemDish;
-import com.lunchtime.repository.MenuItemDishRepository;
 
-@Service
-public class MenuItemDishService {
 
-    private final MenuItemDishRepository menuItemDishRepository;
 
-    public MenuItemDishService(MenuItemDishRepository menuItemDishRepository) {
-        this.menuItemDishRepository = menuItemDishRepository;
-    }
+public interface MenuItemDishService {
 
-    public MenuItemDish save(MenuItemDish menuItemDish) {
-        return menuItemDishRepository.save(menuItemDish);
-    }
+    public MenuItemDish save(MenuItemDish menuItemDish);
 
-    public Page<MenuItemDish> findAll(Pageable pageable) {
-        return menuItemDishRepository.findAll(pageable);
-    }
+    public Page<MenuItemDish> findAll(Pageable pageable);
 
-    public Optional<MenuItemDish> findById(Long id) {
-        return menuItemDishRepository.findById(id);
-    }
+    public Optional<MenuItemDish> findById(Long id);
 }
